# Covenant/API/Models/CommunicationType.cs
// <auto-generated>
// Code generated by Microsoft (R) AutoRest Code Generator.
// Changes may cause incorrect behavior and will be lost if the code is
// regenerated.
// </auto-generated>

namespace LemonSqueezy.API.Models
{
    using Newtonsoft.Json;
    using Newtonsoft.Json.Converters;
    using System.Runtime;
    using System.Runtime.Serialization;

    /// <summary>
    /// Defines values for CommunicationType.
    /// </summary>
    [JsonConverter(typeof(StringEnumConverter))]
    public enum CommunicationType
    {
        [EnumMember(Value = "HTTP")]
        HTTP,
        [EnumMember(Value = "SMB")]
        SMB,
        [EnumMember(Value = "Bridge")]
        Bridge
    }
    internal static class CommunicationTypeEnumExtension
    {
        internal static string ToSerializedValue(this CommunicationType? value)
        {
            return value == null ? null : ((CommunicationType)value).ToSerializedValue();
        }

        internal static string ToSerializedValue(this CommunicationType value)
        {
            switch( value )
            {
                case CommunicationType.HTTP:
                    return "HTTP";
                case CommunicationType.SMB:
                    return "SMB";
                case CommunicationType.Bridge:
                    return "Bridge";
            }
            return null;
        }

        internal static CommunicationType? ParseCommunicationType(this string value)
        {
            switch( value )
            {
                case "HTTP":
                    return CommunicationType.HTTP;
                case "SMB":
                    return CommunicationType.SMB;
                case "Bridge":
                    return CommunicationType.Bridge;
            }
            return null;
        }
    }
}